requirements.txt
SQLAlchemy==1.3.13
selenium==3.141.0
beautifulsoup4==4.8.2
PyMySQL==0.9.3
python-dotenv==0.12.0
